+++ packages/bpk-component-table/src/BpkTableHead.js
@@ -19,7 +19,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-const BpkTableHead = props => <thead>{props.children}</thead>;
+const BpkTableHead = props => <thead {...props} />;
 
 BpkTableHead.propTypes = {
   children: PropTypes.node.isRequired,